+++ tests/sksl/shared/ScopedSymbol.metal
@@ -5,13 +5,13 @@ struct S {
     int i;
 };
 struct Uniforms {
-    half4 colorGreen;
-    half4 colorRed;
+    float4 colorGreen;
+    float4 colorRed;
 };
 struct Inputs {
 };
 struct Outputs {
-    half4 sk_FragColor [[color(0)]];
+    float4 sk_FragColor [[color(0)]];
 };
 struct Globals {
     int glob;
